learning graphql - this is bad code
In some ways this is definitely regressive. Is it ugly – yes, extremely!
But does it get all the nodes (from all 5 pages) and return them in a
single array? Yes.

Does it swallow errors if any should be encountered? Almost certainly
also yes. Could it stay like this for a good long while before anyone
notices? ...yes, yes it could!